the problem is the preview component (something laggy when I resize it)
