+++ client/src/App.tsx
@@ -7,19 +7,17 @@ const Bundle = lazy(() => import('./pages/bundle'));
 
 const App = () => {
   return (
-    <div style={{ height: '100vh' }}>
+    <>
       <BrowserRouter>
         <Header />
-        <div style={{ height: '92vh' }}>
-          <Suspense fallback={<h1>Loading...</h1>}>
-            <Switch>
-              <Route path="/bundles/:id" component={Bundle} />
-              <Route path="/" component={Home} />
-            </Switch>
-          </Suspense>
-        </div>
+        <Suspense fallback={<h1>Loading...</h1>}>
+          <Switch>
+            <Route path="/bundles/:id" component={Bundle} />
+            <Route path="/" component={Home} />
+          </Switch>
+        </Suspense>
       </BrowserRouter>
-    </div>
+    </>
   );
 };
