+++ client/src/components/code-editor.module.sass
@@ -2,8 +2,8 @@
 
 .wrapper
   position: relative
-  width: 100%
   height: 100%
+  width: calc(100% - 10px)
   &:hover .format
     opacity: 1
