+++ client/src/components/preview.module.sass
@@ -1,5 +1,4 @@
 .wrapper
-  // background-color: red
   position: relative
   height: 100%
   flex-grow: 1
