+++ client/src/components/resizable.tsx
@@ -9,6 +9,7 @@ interface ResizableProps {
 const Resizable: React.FC<ResizableProps> = ({ direction, children }) => {
   const [innerHeight, setInnerHeight] = useState(window.innerHeight);
   const [innerWidth, setInnerWidth] = useState(window.innerWidth);
+  // 75% of the initial window width
   const [width, setWidth] = useState(window.innerWidth * 0.75);
 
   let resizableProps: ResizableBoxProps;
@@ -21,10 +22,13 @@ const Resizable: React.FC<ResizableProps> = ({ direction, children }) => {
         window.clearTimeout(timer);
       }
 
+      // resize it smoother
       timer = window.setTimeout(() => {
         setInnerHeight(window.innerHeight);
         setInnerWidth(window.innerWidth);
 
+        // If 75% of the current window width is less than the width of the element
+        // adjust the width of the element to 75% of the window width
         if (window.innerWidth * 0.75 < width) {
           setWidth(window.innerWidth * 0.75);
         }
@@ -41,7 +45,9 @@ const Resizable: React.FC<ResizableProps> = ({ direction, children }) => {
   if (direction === 'horizontal') {
     resizableProps = {
       className: 'resize-horizontal',
-      minConstraints: [innerWidth * 0.2, Infinity],
+      // Take at least 20% of the width and 100% of height
+      minConstraints: [innerWidth * 0.4, Infinity],
+      // Take as maximum value 75% of the width and 100% of height
       maxConstraints: [innerWidth * 0.75, Infinity],
       height: Infinity,
       width,
@@ -52,7 +58,9 @@ const Resizable: React.FC<ResizableProps> = ({ direction, children }) => {
     };
   } else {
     resizableProps = {
+      // Take at least 100% of the width and 24px of height
       minConstraints: [Infinity, 24],
+      // Take as maximum value 100% of the width and 90% of height
       maxConstraints: [Infinity, innerHeight * 0.9],
       height: 300,
       width: Infinity,
