+++ client/src/index.sass
@@ -1,3 +1,5 @@
+@import "https://cdn.jsdelivr.net/npm/bulma@0.9.2/css/bulma.min.css"
+
 *
   box-sizing: border-box
