--- package.json
@@ -1,5 +1,0 @@
-{
-  "devDependencies": {
-    "http-proxy": "^1.18.1"
-  }
-}
